backend/requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.0.0
numpy>=1.24.0

# ML  (CPU-only — no CUDA needed)
ultralytics>=8.3.0
# On ARM Linux (Oracle), install CPU-only PyTorch:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.1.0
torchvision>=0.16.0
